+++ src/views/dashboard/DashboardChart.vue
@@ -1,12 +1,12 @@
 <template>
-  <TopBodyCardComponent :title="caption">
+  <TopBodyCardComponent :title="chartLabel">
     <template v-slot:content>
       <v-layout fill-height>
         <ChartComponent
           id="dashboard-chart"
-          type="doughnut"
+          type="pie"
           :dataset="chartData"
-          :options="options"
+          :options="chartOptions"
         />
       </v-layout>
     </template>
@@ -17,7 +17,7 @@
 import ChartComponent from '@/components/chart/ChartComponent';
 import TopBodyCardComponent from '@/components/TopBodyCardComponent.vue';
 
-import tokensChartBuilder from '@/utils/tokensChartBuilder';
+import tokensChartHelper from '@/utils/tokensChartHelper';
 import { mapGetters } from 'vuex';
 
 export default {
@@ -26,9 +26,6 @@ export default {
     ChartComponent,
     TopBodyCardComponent,
   },
-  data: () => ({
-    allData: null,
-  }),
   computed: {
     ...mapGetters('spreadsheet', {
       abrTokens: 'abrTokens',
@@ -39,70 +36,21 @@ export default {
       pool: 'pool',
       tokens: 'tokens',
     }),
-    caption() {
-      return this.allData ? `Total ${this.allData.all.label} COM` : '';
-    },
     chartData() {
-      return this.allData
-        ? {
-            labels: [
-              `Bonded ${this.allData.bonded.percent}`,
-              `Not Bonded ${this.allData.unbonded.percent}`,
-              `Unrelease rewards ${this.allData.unreleasedRewards.percent}`,
-              `Burned ${this.allData.burned.percent}`,
-            ],
-            datasets: [
-              {
-                data: [
-                  this.allData.bonded.decimal,
-                  this.allData.unbonded.decimal,
-                  this.allData.unreleasedRewards.decimal,
-                  this.allData.burned.decimal,
-                ],
-                backgroundColor: [
-                  'rgb(47, 157, 119)',
-                  'rgb(238, 51, 0)',
-                  'rgb(255, 102, 0)',
-                  'rgb(0, 0, 0)',
-                ],
-                borderColor: [
-                  'rgb(235, 249, 244)',
-                  'rgb(255, 235, 230)',
-                  'rgb(255, 240, 230)',
-                  'rgb(242, 242, 242)',
-                ],
-              },
-            ],
-          }
-        : {};
+      return tokensChartHelper.getChartData({
+        abrTokens: this.abrTokens,
+        params: this.params,
+        pool: this.pool,
+        tokens: this.tokens,
+        vbrTokens: this.vbrTokens,
+      });
     },
-    options() {
-      return {
-        responsive: true,
-        maintainAspectRatio: false,
-        plugins: {
-          legend: {
-            display: false,
-          },
-          tooltip: {
-            callbacks: {
-              label: function(tooltipItem) {
-                return tooltipItem.label;
-              },
-            },
-          },
-        },
-      };
+    chartLabel() {
+      return tokensChartHelper.getChartLabel();
     },
-  },
-  created() {
-    this.allData = tokensChartBuilder.build({
-      abrTokens: this.abrTokens,
-      params: this.params,
-      pool: this.pool,
-      tokens: this.tokens,
-      vbrTokens: this.vbrTokens,
-    });
+    chartOptions() {
+      return tokensChartHelper.getChartOptions();
+    },
   },
 };
 </script>